moved tool in test-util.go to promutil/util.go
