+++ npm/iptm/iptm_test.go
@@ -159,7 +159,7 @@ func TestAdd(t *testing.T) {
 
 	newGaugeVal, err3 := promutil.GetValue(metrics.NumIPTableRules)
 	newCountVal, err4 := promutil.GetCountValue(metrics.AddIPTableRuleExecTime)
-	metrics.NotifyIfErrors(t, err1, err2, err3, err4)
+	promutil.NotifyIfErrors(t, err1, err2, err3, err4)
 	if newGaugeVal != gaugeVal+1 {
 		t.Errorf("Change in iptable rule number didn't register in prometheus")
 	}
@@ -198,7 +198,7 @@ func TestDelete(t *testing.T) {
 	}
 
 	newGaugeVal, err2 := promutil.GetValue(metrics.NumIPTableRules)
-	metrics.NotifyIfErrors(t, err1, err2)
+	promutil.NotifyIfErrors(t, err1, err2)
 	if newGaugeVal != gaugeVal-1 {
 		t.Errorf("Change in iptable rule number didn't register in prometheus")
 	}
