--- npm/metrics/test-util.go
@@ -1,22 +1,0 @@
-package metrics
-
-import "testing"
-
-// NotifyIfErrors writes any non-nil errors to the testing utility
-func NotifyIfErrors(t *testing.T, errors ...error) {
-	allGood := true
-	for _, err := range errors {
-		if err != nil {
-			allGood = false
-			break
-		}
-	}
-	if !allGood {
-		t.Errorf("Encountered these errors while getting metric values: ")
-		for _, err := range errors {
-			if err != nil {
-				t.Errorf("%v", err)
-			}
-		}
-	}
-}
